pyproject.toml
[project]
name = "cheveningbrew-auth-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.11",
    "google-auth[pyopenssl]>=2.38.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
